Removing old swift files
